backend/requirements.txt
# ── Web framework ──────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0

# ── Backboard.io multi-agent SDK (legacy layer) ─────────────────────────
backboard-sdk>=1.5.3

# ── Google Gemini SDK (new intelligence layer) ──────────────────────────
google-genai>=1.0.0

# ── HTTP client (all external API calls) ────────────────────────────────
httpx>=0.27.0

# ── Retry + resilience ───────────────────────────────────────────────────
tenacity>=8.3.0

# ── Environment & config ─────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Google Sheets ingestion (optional — uncomment for server-side fetch) ──
# google-api-python-client>=2.130.0
# google-auth>=2.29.0
